requirements.txt
# AI Log Analyzer Dependencies

# Core
requests>=2.31.0        # HTTP client for LLM API
pyyaml>=6.0.1           # YAML config parsing
python-dotenv>=1.0.0    # .env file loading
